Move scheduler API into another app
