+++ easyspider/urls.py
@@ -13,12 +13,7 @@ urlpatterns = [
          {'status': 'finished'}, name='stop-task'),
     path('node/list/', views.NodeListView.as_view(), name='list-node'),
     path('node/modify/', views.modify_node, name='modify-node'),
+    path('node/set-status/<int:node_id>/<int:status>/', views.set_node_status, name='set-node-status'),
     path('node/create/', views.create_node, name='create-node'),
     path('node/push/', views.push_node_template, name='push-node'),
-    path('node/list-online/', views.list_online_node, name='list-online-node'),
-    path('node/get/<int:node_id>/', views.get_node_by_id, name='get-node-by-id'),
-    path('node/set-status/<int:node_id>/<int:status>/', views.set_node_status, name='set-node-status'),
-    path('job/list/<int:status>/', views.list_job_by_status, name='job-list-by-status'),
-    path('job/set-status/<str:job_id>/<int:status>/', views.set_job_status, name='set-job-status'),
-    path('job/set-node/<str:job_id>/<int:node_id>/', views.set_job_node, name='set-job-node'),
 ]
